Added in power up
Currently it's just a sphere in the islandlevel scene, but it does dissapear on contact with player.
Later i plan on making it so that it's randomly generated on the map every couple of seconds.
There's also a section commented out in the code in relation to particle effects - which we can add in later.