Fixed projectiles damaging shooter. Round functionality. etc.
I added the EGO for both players as suggested by Aaron in class and now the projectiles don't damage the player firing them. I also made the projectiles destroy themselves in a certain amount time and upon contact with the other player. I added in a round counter in the components (nothing on UI yet) but I need to fix it so that it only counts towards the player winning the current round.