Fiddling with player input scripts and added camera control
I fixed the health bar for player 2 to function the same as player 1's health bar.  I was trying to make the round counter work and have custom text that can be changed through the component so it's easier to display the class name in ui. I also added camera control based on the unity tanks tutorial so now the camera moves with both players.