Fiddling with scripts
Changed name text into a string so that each class is now appropriately labeled on the UI. Added in a button to return to the menu but still working on trying to have it enabled upon player victory. Also added a character select screen with no functionality at the moment.